More flexibility in version byte handling + allowing None as value
As suggested by neonknight on GitHub, we now allow passing None to
validate() as the value for magicbyte (which has been renamed to
'decimal_prefixes' since it's more descriptive imo) to skip version
byte validation (this would allow any Bitcoin-like currency's address
to validate) and set the default to (0, 5) to make it easier to
use the function out-of-the-box (so by default it now validates all
non-testnet Bitcoin addresses). Use of a single integer is still
supported. Also updated comment to reflect changes + provide more
info about the version bytes.